BUGFIX: Error 404 Solution
This is a simple solution for the error 404 in [img] nnCode that breaks chatlist autoscroll in chromium based nrowsers. it defaults to the delete png because it seemed like as good a choice as any but I wonder if a dedicated image or the warning.ong emoticon might be better?

This issue affects AJAX-Chat 0.8.7, its current branches and all prior versions but is only presented here because I'm unsure ofteh proper way to handle 0.8.7 master merges.